requirements.txt

google-cloud-bigquery
langchain-google-vertexai
langchain-community
langchain-core
langchain-openai
